chore: Refactoring --WIP
