+++ components/Cards/card.module.scss
@@ -104,8 +104,8 @@
     }
   }
   &:hover {
-    // box-shadow: 2px 2px 0px $color-dark;
-    // transform: translate(-2px, -2px);
+    box-shadow: 2px 2px 0px $color-dark;
+    transform: translate(-2px, -2px);
     background-color: $color-secondary-less;
     color: $color-dark;
     font-weight: 900;
