+++ components/Post/PostLayout.js
@@ -1,9 +1,9 @@
 import Head from 'next/head';
 import styles, { post_container } from './layout.module.scss';
 import Link from 'next/link';
-import Header from '../Header/header';
+import Header from '../Header/Header';
 import PostFooter from './PostFooter';
-import Footer from '../Footer/footer';
+import Footer from '../Footer/Footer';
 
 const name = 'Timothy Merritt';
 export const siteTitle = 'timmybytes';
